Extract LandingPage + dialog components, drop ~308 lines from Index.tsx
Index.tsx's no-data branch was 322 lines of mostly-static JSX (header,
file import, sample loader, three large content dialogs, footer links).
Pulled it out into focused components:

  src/components/SupportedFilesDialog.tsx (78 lines)
    Self-contained dialog listing the 9 supported file formats. Data
    moved to module-level PRIMARY_FORMATS / SECONDARY_FORMATS arrays;
    the 9 hand-repeated <div> blocks collapse into 2 short loops.

  src/components/AboutDialog.tsx (90 lines)
    Self-contained dialog with the 4 about sections and 21-item feature
    grid. Sections + features pulled to module-level constants;
    rendering is now two small loops.

  src/components/CreditsDialog.tsx (58 lines)
    Self-contained credits dialog. 16-entry credits list moved to a
    module-level tuple array.

  src/components/LandingPage.tsx (155 lines)
    The full no-data view: header (with the three dialogs and Sponsor
    button), the SEO h1, FileImport, sample loader, BrowserCompat link,
    GitHub link bar (pulled to a constant), and footer (privacy /
    contact / credits / optional admin). Owns the navigate() call for
    the admin button so Index.tsx no longer needs useNavigate.

Props the LandingPage takes (7): onDataLoaded, onOpenFileManager,
autoSave, autoSaveFile, onLoadSample, isLoadingSample, enableAdmin.
Everything else is internal.

Index.tsx changes:
  - No-data branch went from 322 lines to ~16 lines (one LandingPage
    call wrapped in DeviceProvider + the existing FileManagerDrawer).
  - File total: 873 -> 564 lines (-309, -35%).
  - Imports cleaned up: dropped 16 now-unused names (FileText, Info,
    Heart, Github, BookOpen, Shield, Link, Loader2, ExternalLink,
    Download, Dialog primitives, ContactDialog, BrowserCompatDialog,
    FileImport, LocalWeatherDialog) and the unused useNavigate.

Behavior preserved — no JSX content changed, just moved. Dialogs are
still mounted via Radix's <Dialog> primitive (no open/close state to
preserve), so the user-visible behavior is unchanged.

Verified:
  - npm run lint clean (0 problems)
  - tsc --noEmit clean
  - npm run test:run (203 passed)
  - npm run build clean

Remaining on the Index.tsx decomposition roadmap:
  - useDataLoader() hook for the 90-line handleDataLoaded orchestration.
    That brings Index.tsx down to roughly the 350-line target.
